Move more test from use-deeplink
